Push "use client" to the leaves on /ai
Keep static sections on the server, extract the hero canvas, type demo
icons against the presets, and hide validation errors on auto-run like
the homepage editor.
