+++ src/app/(main)/ai/components/hero.tsx
@@ -1,11 +1,10 @@
-"use client"
-
-import { useEffect, useRef } from "react"
 import { StripesDecoration } from "@/app/conf/_design-system/stripes-decoration"
 import { Button } from "@/app/conf/_design-system/button"
 import CheckIcon from "@/app/conf/_design-system/pixelarticons/check.svg?svgr"
 import ArrowDownIcon from "@/app/conf/_design-system/pixelarticons/arrow-down.svg?svgr"
 import blurBean from "@/components/blog-page/blur-bean.webp"
+
+import { SchemaGrid } from "./hero-schema-grid"
 
 const highlights = [
   "Self-describing schemas let agents discover your API",
@@ -91,145 +90,5 @@ export function Hero() {
         </div>
       </div>
     </div>
-  )
-}
-
-function SchemaGrid() {
-  const canvasRef = useRef<HTMLCanvasElement>(null)
-
-  useEffect(() => {
-    const canvas = canvasRef.current
-    if (!canvas) return
-    const ctx = canvas.getContext("2d")
-    if (!ctx) return
-
-    let animId: number
-    let time = 0
-    const nodes: { x: number; y: number; r: number; vx: number; vy: number }[] =
-      []
-
-    function resize() {
-      if (!canvas) return
-      canvas.width = canvas.offsetWidth * (window.devicePixelRatio || 1)
-      canvas.height = canvas.offsetHeight * (window.devicePixelRatio || 1)
-      ctx!.setTransform(1, 0, 0, 1, 0, 0)
-      ctx!.scale(window.devicePixelRatio || 1, window.devicePixelRatio || 1)
-    }
-
-    function init() {
-      resize()
-      const w = canvas!.offsetWidth
-      const h = canvas!.offsetHeight
-      nodes.length = 0
-      for (let i = 0; i < 30; i++) {
-        nodes.push({
-          x: Math.random() * w,
-          y: Math.random() * h,
-          r: Math.random() * 2 + 1,
-          vx: (Math.random() - 0.5) * 0.3,
-          vy: (Math.random() - 0.5) * 0.3,
-        })
-      }
-    }
-
-    function draw() {
-      if (!canvas || !ctx) return
-      const w = canvas.offsetWidth
-      const h = canvas.offsetHeight
-
-      ctx.clearRect(0, 0, w, h)
-      time += 0.005
-
-      for (const n of nodes) {
-        n.x += n.vx
-        n.y += n.vy + Math.sin(time + n.x * 0.02) * 0.1
-        if (n.x < 0) n.x = w
-        if (n.x > w) n.x = 0
-        if (n.y < 0) n.y = h
-        if (n.y > h) n.y = 0
-      }
-
-      const maxDist = 150
-      for (let i = 0; i < nodes.length; i++) {
-        for (let j = i + 1; j < nodes.length; j++) {
-          const dx = nodes[i].x - nodes[j].x
-          const dy = nodes[i].y - nodes[j].y
-          const dist = Math.sqrt(dx * dx + dy * dy)
-          if (dist < maxDist) {
-            ctx.beginPath()
-            ctx.moveTo(nodes[i].x, nodes[i].y)
-            ctx.lineTo(nodes[j].x, nodes[j].y)
-            const alpha = (1 - dist / maxDist) * 0.08
-            ctx.strokeStyle = `rgba(255,204,239,${alpha})`
-            ctx.stroke()
-          }
-        }
-      }
-
-      for (const n of nodes) {
-        ctx.beginPath()
-        ctx.arc(n.x, n.y, n.r, 0, Math.PI * 2)
-        ctx.fillStyle = "rgba(255,204,239,0.15)"
-        ctx.fill()
-      }
-    }
-
-    function frame() {
-      draw()
-      animId = requestAnimationFrame(frame)
-    }
-
-    const stillness = window.matchMedia("(prefers-reduced-motion: reduce)")
-
-    let running = false
-    let onScreen = false
-
-    function sync() {
-      const shouldRun = onScreen && !document.hidden && !stillness.matches
-      if (shouldRun === running) return
-      running = shouldRun
-      if (shouldRun) frame()
-      else cancelAnimationFrame(animId)
-    }
-
-    function onResize() {
-      resize()
-      const w = canvas!.offsetWidth
-      const h = canvas!.offsetHeight
-      for (const n of nodes) {
-        n.x = Math.min(n.x, w)
-        n.y = Math.min(n.y, h)
-      }
-      if (!running) draw()
-    }
-
-    const observer = new IntersectionObserver(([entry]) => {
-      onScreen = entry.isIntersecting
-      sync()
-    })
-
-    init()
-    draw()
-    observer.observe(canvas)
-    stillness.addEventListener("change", sync)
-    document.addEventListener("visibilitychange", sync)
-    window.addEventListener("resize", onResize)
-
-    return () => {
-      running = false
-      cancelAnimationFrame(animId)
-      observer.disconnect()
-      stillness.removeEventListener("change", sync)
-      document.removeEventListener("visibilitychange", sync)
-      window.removeEventListener("resize", onResize)
-    }
-  }, [])
-
-  return (
-    <canvas
-      ref={canvasRef}
-      aria-hidden
-      className="pointer-events-none absolute inset-0 z-0 opacity-60"
-    />
   )
 }
